+++ mod-source-record-storage-server/src/main/java/org/folio/config/ApplicationConfig.java
@@ -1,6 +1,5 @@
 package org.folio.config;
 
-import io.vertx.core.Vertx;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Value;
@@ -9,8 +8,6 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 import org.folio.kafka.KafkaConfig;
-import org.folio.kafka.cache.KafkaInternalCache;
-import org.folio.kafka.cache.util.CacheUtil;
 
 @Configuration
 @ComponentScan(basePackages = {
@@ -49,19 +46,5 @@ public class ApplicationConfig {
     LOGGER.debug("kafkaConfig: {}", kafkaConfig);
 
     return kafkaConfig;
-  }
-
-  @Bean
-  public KafkaInternalCache kafkaInternalCache(KafkaConfig kafkaConfig, Vertx vertx,
-                                               @Value("${srs.kafka.cache.cleanup.interval.ms:3600000}") long cleanupInterval,
-                                               @Value("${srs.kafka.cache.expiration.time.hours:3}") int expirationTime) {
-    KafkaInternalCache kafkaInternalCache = KafkaInternalCache.builder()
-      .kafkaConfig(kafkaConfig)
-      .build();
-
-    kafkaInternalCache.initKafkaCache();
-    CacheUtil.initCacheCleanupPeriodicTask(vertx, kafkaInternalCache, cleanupInterval, expirationTime);
-
-    return kafkaInternalCache;
   }
 }